Always check license when compiling, use new license checker
